+++ DefenceGame/DefenceGame/Golem.cpp
@@ -6,7 +6,7 @@ Golem::Golem(ENTITY_TYPE type, std::string renderString, COLOR color, int hp, in
 	_type = type;
 	_renderString = renderString;
 	_color = color;
-	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / (float)40));
+	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / 40.0f));
 	_moveTime = moveTime;
 	_rewardGold = rewardGold;
 	_originColor = _color;
