+++ DefenceGame/DefenceGame/Ogre.cpp
@@ -7,7 +7,7 @@ Ogre::Ogre(ENTITY_TYPE type, std::string renderString, COLOR color, int hp, int 
 	_type = type;
 	_renderString = renderString;
 	_color = color;
-	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / (float)40));
+	_hp = std::round((float)hp * (float)(1 + GET_SINGLETON(WaveManager)->getCurrentWave() / 40.0f));
 	_moveTime = moveTime;
 	_rewardGold = rewardGold;
 	_originColor = _color;
